\name{usefnptr-package}
\alias{usefnptr-package}
\alias{usefnptr}
\docType{package}
\title{
\packageTitle{usefnptr}
}
\description{
\packageDescription{usefnptr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{usefnptr}
\packageIndices{usefnptr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{usefnptr}

Maintainer: \packageMaintainer{usefnptr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
